fixed(for the most part) bonus stanima from full soda
Through a lot of trial and error, I found out that trying to use a timer in a class is an act of futility as there is no access to the scenetree through get_tree(), and add_child() does not apparently add the timer to the tree. this means the nice clean way of dealing with timed bonuses isn't going to work and has to be brought into where the bonus set.
still, the class has proven a better alternative to the dictionary I was previously using and I may want to move playerInventory into its own class for the planned npc trading mechanic and potential enemy loot mechanic