made "reset" primary in input chain, moved handling all into level
there were issues with reset propogating multiple times and messingg with inventory screen, I removed stats input handling of reset and moved reset handling to the top of process, letting mode make the call on what to do when reset is pressed. On the plus side this allowed me to get rid of some variables of dubious worth.
need to ensure gemna still works with this model, but I think we should be fine